+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Revamped support for `--validate-server` (short `-s`)
+  - Requires the use of server base path(s) on the client.
+  - Localhost is now always allowed on all known base paths.
+  - Support for proxy headers (e.g. `X-Forwarded-Host`).
 - Better support for resolving relative path references.
 - Be more resilient to parser panics when using `--watch`
 - Update Docker build to use Go 1.12 and Go modules.